+++ R/normexp_mgcv.R
@@ -138,7 +138,7 @@ normexp <- function(link = list("identity", "log", "log"), s = -1){
     s       <- family$s
     ##Define parameters
 
-    l0<-dnormexp(x=y, mu=mu, sigma_v=sigma_v, lambda=lambda, s=s, deriv=4, tri=family$tri, log.p = TRUE)
+    l0<-dnormexp(x=y, mu=mu, sigma_v=sigma_v, lambda=lambda, s=s, deriv=4, tri=family$tri, log.p = TRUE, check=FALSE)
 
     l<-sum(l0)
 
@@ -213,10 +213,13 @@ normexp <- function(link = list("identity", "log", "log"), s = -1){
       qr.x_all<-qr(x[,-which(duplicated(colnames(x))), drop=FALSE])
       s<-family$s
 
+      # para<-reparametrize(mean = qr.fitted(qr.x_all, y),
+      #                         sd = qr.fitted(qr.x_all, abs(qr.resid(qr.x_all, y))),#sqrt(sum(qr.resid(qr.x_all, y)^2)/nrow(x1)),
+      #                       skew = rep((sum((y - mean(y))^3)/length(y))/(sum((y - mean(y))^2)/length(y))^(3/2),length(y)), s=s, family="normexp")
       para<-reparametrize(mean = qr.fitted(qr.x_all, y),
-                              sd = qr.fitted(qr.x_all, abs(qr.resid(qr.x_all, y))),#sqrt(sum(qr.resid(qr.x_all, y)^2)/nrow(x1)),
-                            skew = rep((sum((y - mean(y))^3)/length(y))/(sum((y - mean(y))^2)/length(y))^(3/2),length(y)), s=s, family="normexp")
-
+                          sd = abs(qr.resid(qr.x_all, y)),#sqrt(sum(qr.resid(qr.x_all, y)^2)/nrow(x1)),
+                          skew = rep((sum((y - mean(y))^3)/length(y))/(sum((y - mean(y))^2)/length(y))^(3/2),length(y)),
+                          family="normexp", s=s)
       yt1<-para$mu
 
       # 1) Ridge regression for the location parameter
@@ -253,30 +256,30 @@ normexp <- function(link = list("identity", "log", "log"), s = -1){
   }) ## initialize
 
   #Random number generation for normexp
-  rd <- function(mu, wt, scale, s) {
+  rd <- function(mu, wt, scale) {
     ## random number generation
     mu <- as.matrix(mu)
     if(ncol(mu)==1){ mu <- t(mu) }
 
-    return(rnormexp(n=nrow(mu), mu = mu[,1], sigma_v = mu[,2], lambda = mu[,3], s=s))
+    return(rnormexp(n=nrow(mu), mu = mu[,1], sigma_v = mu[,2], lambda = mu[,3], s=attr(mu,"s")))
   } ## random number generation
 
   #Quantile function of normexp
-  qf <- function(p, mu, wt, scale, s) {
+  qf <- function(p, mu, wt, scale) {
     ##quantile function
     mu <- as.matrix(mu)
     if(ncol(mu)==1){ mu <- t(mu) }
 
-    return(qnormexp(p=p, mu = mu[,1], sigma_v = mu[,2], lambda = mu[,3], s=s))
+    return(qnormexp(p=p, mu = mu[,1], sigma_v = mu[,2], lambda = mu[,3], s=attr(mu,"s")))
   } ##quantile function
 
   #Cumulative distribution function of normexp
-  cdf <- function(q, mu, wt, scale, logp, s) {
+  cdf <- function(q, mu, wt, scale, logp) {
     ##cumulative distribution function
     mu <- as.matrix(mu)
     if(ncol(mu)==1){ mu <- t(mu) }
 
-    return(pnormexp(q=q, mu = mu[,1], sigma_v = mu[,2], lambda = mu[,3], s=s, log.p = logp))
+    return(pnormexp(q=q, mu = mu[,1], sigma_v = mu[,2], lambda = mu[,3], s=attr(mu,"s")))
   } ##cumulative distribution function
 
   #Prediction function
@@ -323,6 +326,11 @@ normexp <- function(link = list("identity", "log", "log"), s = -1){
     }
   } ##predict
 
+  postproc <- expression({
+    attr(object$fitted.values,"s")<-object$family$s
+    object$fitted.values
+  })
+
   structure(list(family="normexp",ll=ll, link=paste(link), nlp=npar,
                  tri = mgcv::trind.generator(npar), ## symmetric indices for accessing derivative arrays
                  initialize=initialize,
@@ -332,6 +340,7 @@ normexp <- function(link = list("identity", "log", "log"), s = -1){
                  qf=qf,
                  cdf=cdf,
                  predict=predict,
+                 postproc=postproc,
                  linfo = stats, ## link information list
                  d2link=1, d3link=1, d4link=1, ## signals to fix.family.link that all done
                  ls=1, ## signals that ls not needed here